samples/hackathons/github-copilot-banking-dev/resources/starter/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic>=2.0,<3.0
sqlalchemy>=2.0,<3.0
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-mock==3.14.0
httpx==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
structlog==24.2.0
ruff==0.4.7
mypy==1.10.0
